deeplearning/zhengfangjiaowu-master/web_app/requirements.txt
keras
tensorflow
numpy
pillow
flask
requests
gevent
bs4